Refactor data prep into its own file, updates tally zone file
Also makes the available years dynamic based on what is available in the tally zone file, which is being updated progressively.
